Code refactoring
- Remove obsolete code for ITIR-lowering to SDFG from dace_iterator module
- Move utilities and workflow-related methods to the dace_fieldview
  module
- The only code left in dace_iterator module is the newly-added dace
  orchestration features, that will require modifications to work with
  GTIR programs.
